Refatoração
Divisão dos controladores e rotas de usuários
