+++ src/server.ts
@@ -1,25 +1,14 @@
 // Importações
-import express, { Request, Response } from "express";
+import express from "express";
 import connect from "../db/connect";
-import * as userController from "./modules/user-controller";
+import routes from "./routes/userRouter";
 
 const app = express();
-const port = process.env.PORT || 3000;
+app.use(express.json());
 connect;
 
-app.use(express.json())
+app.use("/", routes);
 
-app.use("/login", userController.loginControler);
-app.use("/register", userController.registerControler);
-app.use("/update", userController.updateControler);
-
-app.get("/", (req: Request, res: Response) => {
-  try {
-    res.status(200).send("Servidor Online");
-  } catch (err) {
-    res.status(500).send("Erro interno no servidor");
-  }
-});
-app.listen(port, () => {
+app.listen(3000, () => {
   console.log("Servidor iniciado na porta 3000!");
 });
